use `usize` type parameters for `min_values` and `max_values`

+++ token/cli/src/clap_app.rs
@@ -318,8 +318,8 @@ pub fn multisig_signer_arg<'a, 'b>() -> Arg<'a> {
         .value_name("MULTISIG_SIGNER")
         .takes_value(true)
         .multiple(true)
-        .min_values(0u64)
-        .max_values(MAX_SIGNERS as u64)
+        .min_values(0_usize)
+        .max_values(MAX_SIGNERS)
         .help(MULTISIG_SIGNER_ARG.help)
 }
 
@@ -1178,8 +1178,8 @@ pub fn app<'a, 'b>(
                         .takes_value(true)
                         .index(2)
                         .required(true)
-                        .min_values(MIN_SIGNERS as u64)
-                        .max_values(MAX_SIGNERS as u64)
+                        .min_values(MIN_SIGNERS)
+                        .max_values(MAX_SIGNERS)
                         .help(multisig_member_help),
                 )
                 .arg(
@@ -1368,7 +1368,7 @@ pub fn app<'a, 'b>(
                         .value_name("PUBKEY:ROLE")
                         .takes_value(true)
                         .multiple(true)
-                        .min_values(0u64)
+                        .min_values(0_usize)
                         .help("Additional pubkey(s) required for a transfer hook and their \
                             role, in the format \"<PUBKEY>:<ROLE>\". The role must be \
                             \"readonly\", \"writable\". \"readonly-signer\", or \"writable-signer\".\
@@ -2207,7 +2207,7 @@ pub fn app<'a, 'b>(
                         .value_name("ACCOUNT_ADDRESS")
                         .takes_value(true)
                         .multiple(true)
-                        .min_values(0u64)
+                        .min_values(0_usize)
                         .help("The token accounts to withdraw from")
                 )
                 .arg(
